requirements.txt
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
Django==4.2.1
django-channels==0.7.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
zope.interface==6.0
